Fix `vi/vim/vim-minimal` package issues
Turns out Arch deleted the `vim-minimal` package, but `vim` and `gvim`
both "replace" `vim-minimal`.  So, trying to install `vim-minimal`
resulted in reinstalling the `vim` that was just removed.

Instead, ensure that both `vim` and `vim-minimal` (the package, not the
requirement) are absent.  And install `vi`, which peacefully coexists
with `gvim`, to satisfy the requirement of having a tiny `vi`-clone as
one of the very first packages.