Install compatibility shim for ncurses5 API
Found that there was an AUR package to do this. But, because `aura`
isn't compiled until after something needs the ncurses5 API, the old
hack of using a version from `/nix/store/` is still necessary for
bootstrapping.

Updated the bootstrapping version to check for the "real" compat
package, fix the `changed` status, and add explanatory comments
(including a TODO).

Added three tasks in `arch-build` (after `aura` has been installed):

1. Remove the no-longer-needed `Nix` version (needs to be cleared out of
the way before installing the compatibility shim)

2. Install the weird, custom GPG key for the pkg
(It belongs to Thomas Dickey, XTerm/ncurses guy, so I trust it)

3. Install the compatibility package