Remove weird `[initial]` host group
Intent was that a host would be placed in `[initial]`, then
`00-initial.yml` could be included in `site.yml`, but would only apply
when any hosts were in that group.  (But, I never got around to
including it.)

Instead, remove the `[initial]` group, and call the `00-initial.yml`
file directly via `ansible-playbook`.